refactor: section data (#2885)
