+++ next/components/Home/components/Section/components/SectionPublications/common/entities.ts
